Fix bug that's been in the GATK for a very long time: update nReads (as well as nRecords), so that INFO logging doesn't say 'skipped 0 of 0 reads'.  While I'm in there, update TraversalStatistics to store longs.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@2959 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/gatk/traversals/TraversalStatistics.java
@@ -13,13 +13,13 @@ public class TraversalStatistics {
     // Number of records (loci, reads) we've processed
     public static long nRecords;
     // How many reads have we processed, along with those skipped for various reasons
-    public static int nReads;
-    public static int nSkippedReads;
-    public static int nUnmappedReads;
-    public static int nNotPrimary;
-    public static int nBadAlignments;
-    public static int nSkippedIndels;
-    public static int nDuplicates;
+    public static long nReads;
+    public static long nSkippedReads;
+    public static long nUnmappedReads;
+    public static long nNotPrimary;
+    public static long nBadAlignments;
+    public static long nSkippedIndels;
+    public static long nDuplicates;
 
     static {
         reset();
